// frontend/src/api/models/UserDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UserRoleDto } from './UserRoleDto';
export type UserDto = {
    id?: number;
    yandexId?: string;
    name?: string;
    email?: string;
    registrationStatus?: string;
    roles?: Array<UserRoleDto>;
};